save dev state
